requirements.txt
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cross-web==0.4.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
graphql-core==3.2.7
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
motor==3.7.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
openai==2.15.0
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pypdf==6.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
redis==6.4.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
strawberry-graphql==0.289.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.5.0
websockets==16.0
